Refactored serialdata.c. Fixed bug stopping parsing!
Parsing wasn't allowing a ' ' character.
