+++ gui.h
@@ -156,8 +156,7 @@ private:
 	void showMessageInLog(const char *message);
 	void LogImuData(ImuData imuData);
 	void LogOrientationData(YawPitchRoll orientation);
-		
-		
+				
 	// Build UI components
 	void BuildMenu();
 	wxBoxSizer* BuildLeftPanel(wxPanel *panel);
